DLCs_Main: Split the Halloween DLC into a separate file.
