+++ DLCs_Main/Halloween.h
@@ -1,0 +1,11 @@
+// Halloween DLC functions.
+#pragma once
+
+#include <SADXStructs.h>
+extern const PVMEntry HalloweenTextures;
+
+void Halloween_Load(ObjectMaster *a1);
+void LoadHalloweenStuff_StationSquare(void);
+void LoadHalloweenStuff_Twinkle1(void);
+void LoadHalloweenStuff_Twinkle2(void);
+void LoadHalloweenStuff_Twinkle3(void);
